+++ src/os2datascanner/projects/report/tests/test_match_web_alias_relations_using_mainpage_view.py
@@ -108,7 +108,7 @@
 #     progress=None)
 
 # egon_positive_match = messages.MatchesMessage(
-#     scan_spec=egon_scan_spec._replace(scan_tag=scan_tag0),
+#     scan_spec=messages.replace(egon_scan_spec, scan_tag=scan_tag0),
 #     handle=egon_web_handle,
 #     matched=True,
 #     matches=[messages.MatchFragment(
@@ -117,7 +117,7 @@
 # )
 
 # egon_positive_match_1 = messages.MatchesMessage(
-#     scan_spec=egon_scan_spec._replace(scan_tag=scan_tag1),
+#     scan_spec=messages.replace(egon_scan_spec, scan_tag=scan_tag1),
 #     handle=egon_web_handle_1,
 #     matched=True,
 #     matches=[messages.MatchFragment(
@@ -127,7 +127,7 @@
 
 
 # egon_positive_match_2 = messages.MatchesMessage(
-#     scan_spec=egon_scan_spec._replace(scan_tag=scan_tag2),
+#     scan_spec=messages.replace(egon_scan_spec, scan_tag=scan_tag2),
 #     handle=egon_web_handle_2,
 #     matched=True,
 #     matches=[messages.MatchFragment(
@@ -175,7 +175,7 @@
 #     progress=None)
 
 # kjeld_positive_match = messages.MatchesMessage(
-#     scan_spec=kjeld_scan_spec._replace(scan_tag=scan_tag0),
+#     scan_spec=messages.replace(kjeld_scan_spec, scan_tag=scan_tag0),
 #     handle=kjeld_web_handle,
 #     matched=True,
 #     matches=[messages.MatchFragment(